[refactor] #91 - grade 위치 변경에 따른 변경
