+++ src/main/java/gitbal/backend/domain/user/User.java
@@ -2,7 +2,7 @@ package gitbal.backend.domain.user;
 
 
 import gitbal.backend.domain.school.School;
-import gitbal.backend.domain.grade.Grade;
+import gitbal.backend.global.constant.Grade;
 import gitbal.backend.domain.majorlanguage.MajorLanguage;
 import gitbal.backend.domain.onedaycommit.OneDayCommit;
 import gitbal.backend.domain.region.Region;
